Add ImageRecord type in database.go
The commit includes an update to the CHANGELOG.md - it now states that we are copying the API description from the private api_service instead of adding it. Furthermore, a new type named ImageRecord has been added to the database.go file. This new type extends gorm.Model and includes fields for DateTimeStart, OrderRecordID, OrderRecord, Image, and Note.

Took 1 minute